Bridge/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
paho-mqtt==2.0.0
pyserial==3.5
requests==2.31.0
urllib3==2.2.0
